# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CircleTrade Agent Stack API
  description: |
    The Circle Agent Stack API exposes CircleTrade AI's export intelligence pipeline as
    individually callable micro-agent endpoints. Each endpoint maps to a distinct intelligence
    module: DDTRS (compliance & HS classification), MeridianFlow (route optimisation),
    and TradeMatch (market & tariff intelligence).

    **Authentication**: All `/api/*` endpoints require an `X-API-Key` header.
    Demo key: `ct-demo-key-2026`
  version: 1.0.0
  contact:
    name: CircleTrade AI
    url: https://circletrade.ai
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: product_description is required
        timestamp:
          type: string
          format: date-time
...
    RouteResult:
      type: object
      properties:
        name: { type: string }
        score: { type: integer }
        transit_days: { type: integer }
        cost_index: { type: number }
        port_efficiency: { type: integer }
        political_risk: { type: string }
        regulatory_alignment: { type: string }
        primary_ports:
          type: array
          items: { type: string }
        recommendation: { type: string }
    RouteResponse:
      type: object
      properties:
        request_id: { type: string }
        query:
          type: object
          properties:
            origin: { type: string }
            destination: { type: string }
            product_category: { type: string, nullable: true }
        routes:
          type: array
          items:
            $ref: '#/components/schemas/RouteResult'
        best_route:
          $ref: '#/components/schemas/RouteResult'
        total_routes_found: { type: integer }
        source: { type: string }
        timestamp: { type: string, format: date-time }
...
    ComplianceResponse:
      type: object
      properties:
        request_id: { type: string }
        passed: { type: boolean }
        hs_code: { type: string }
        category: { type: string }
        destination: { type: string }
        sanctioned: { type: boolean }
        license_required: { type: boolean }
        license_type: { type: string, nullable: true }
        dual_use: { type: boolean }
        check_body: { type: string, nullable: true }
        cert_required: { type: string, nullable: true }
        risk_score: { type: integer }
        risk_level: { type: string, enum: [CRITICAL, HIGH, MODERATE, LOW] }
        warnings:
          type: array
          items: { type: string }
        issues:
          type: array
          items: { type: string }
        recommendation: { type: string }
        source: { type: string }
        timestamp: { type: string, format: date-time }
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/route:
    post:
      summary: Optimise trade corridor / shipping route
      tags: [MeridianFlow]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  type: string
                  example: UK
                destination:
                  type: string
                  example: UAE
                product_category:
                  type: string
                  example: textiles
      responses:
        '200':
          description: Route optimisation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/compliance:
    post:
      summary: Run compliance check for a product and destination
      tags: [DDTRS]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_description, destination, category]
              properties:
                product_description:
                  type: string
                  example: "ultrasound medical imaging device"
                destination:
                  type: string
                  example: India
                category:
                  type: string
                  example: medical
      responses:
        '200':
          description: Compliance check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...